Make int to size_t conversion explicit in loop initialization

+++ AampMPDParseHelper.cpp
@@ -1239,7 +1239,7 @@ double AampMPDParseHelper::GetPeriodDurationFromStart(int &periodIndex)
 	if (!periodStartStr.empty())
 	{
 		double periodStart = ParseISO8601Duration(periodStartStr.c_str());
-		for (size_t p = periodIndex + 1; p < periods.size(); p++)
+		for (size_t p = static_cast<size_t>(periodIndex) + 1; p < periods.size(); p++)
 		{
 			std::string nextPeriodStartStr = periods.at(p)->GetStart();
 			bool hasSegments = aamp_HasSegmentTimeAndSegments(periods.at(p));
